Add dummy data

+++ db/migrate/20171105022645_create_transactions.rb
@@ -1,7 +1,7 @@
 class CreateTransactions < ActiveRecord::Migration[5.1]
   def change
     create_table :transactions do |t|
-      t.float :unix_time
+      t.integer :unix_time
       t.float :value_krw
       t.float :value_btc
